Bump to Ignition v2.0.0-alpha
Go back to git master for ignition-dracut and import Ignition from git
tag v2.0.0-alpha. This required adapting the build to using go modules
by building with `-mod=vendor` and `GO111MODULE=on` and dropping the
`GOPATH` hacks.

We also drop `parsedeps.go` since Ignition moved away from glide.
Instead introduce a new tiny `go-mods-to-bundled-provides.py` script
that just parses `go.mod` and `vendor/modules.txt` to generate virtual
`Provides`.  The `parsedeps.go` approach also had a few issues that are
fixed in the new script. Notably, it would only pick up on subpackages
of dependent repos, and not the main packages (e.g. it skipped over
`ajeddeloh/go-json`). And also, it was generating `Provides` at the
version of the Ignition package (using `%{version}`), instead of the
version of the bundled package itself.