Switch -validate-nonlinux to noarch; move files to /usr/share/ignition
We shouldn't put non-Linux binaries in /usr/bin.  They're opaque data
as far as Linux is concerned, and we can build them on any arch.
Switch the subpackage to noarch, move the binaries to
/usr/share/ignition, and make them non-executable.

Rename files to include architecture triples, and in the Windows case,
the necessary .exe suffix.  These names match the ones used for FCCT
binaries.  For macOS the name now matches the actual name used on the
GitHub releases page; for Windows this is a change.